Modified calc_chi2
calc_chi2 is now generalised for any histogram. The flag for the smeared
case and for the smeared histogram have been removed
TODO - Need to check the variable scalefactor to see if functionality
for this needs to be added in somewhere. Seems that h_smc previously was
scaled using this scale factor, although I can't see where the factor
was defined anywhere.