Uncommented plotting regions
TODO move all of this to separate functions
Note that the syntax for calling fitting.c uses two different energy
regions. The MC energy region is ~700 keV higher than the corresponding
region from the Ge detector. Expect that this is due to losses. The
value in keV should correspond to which MC files should be used and the
value in MeV should describe where the peak is present in the MC.